Handle closing/opening of LCU
Makes LCU Explorer more robust in the face of opening/closing/reopening
of the LCU client, and slightly faster. FE keeps all the state in
`useState` calls which allow the Swagger UI, OpenAPI spec JSON, and the
LCU credentials to load (or unload/reload) in any order. A basic plugin
updates the Swagger UI on changes.

When not connected to the LCU the endpoints still show up but the "Try
It Out" buttons are hidden.